Modification de std, correction de bugs

+++ cpp_lib/basics.cpp
@@ -18,12 +18,12 @@ extern "C" w_variable *et(std::vector<w_variable *> args, variable_table variabl
     }
 
 
-    int a = *(int *)(args[0]->content);
-    int b = *(int *)(args[1]->content);
+    int64_t a = *(int64_t *)(args[0]->content);
+    int64_t b = *(int64_t *)(args[1]->content);
 
     w_variable *r = new w_variable();
     r->type = 2; // int
-    r->content = (void *)(new int(a && b));
+    r->content = (void *)(new int64_t(a && b));
     return r;
 }
